third_party/blink/web_tests/external/wpt/web-animations/interfaces/Animatable/getAnimations-pseudoElement-option-on-view-transition-expected.txt

This is a testharness.js-based test.
[FAIL] Returns view transition animations with pseudoElement option
  assert_equals: Returns view transition animation when pseudoElement is specified with pseudo expected 1 but got 0
Harness: the test ran to completion.
